refactor(pty): collapse duplicated PTY and cmdline helpers into common/ptyexec
internal/agent/loginpty and internal/agentctl/server/process each carried a
byte-identical copy of the Win32 command-line quoting (scanArgChars,
appendEscapedBytes, escapeArg, buildCmdLine) and the PTY wrappers
(PtyHandle, startPTYWithSize, resolveConPtyCmdLine, isBatchScript). The
copies had already drifted: divergent signatures (uint16 vs int), different
error strings, and comments describing different lifecycles.

Both now share internal/common/ptyexec. internal/common is the neutral tier
because agentctl ships as a standalone binary uploaded into containers, so
neither tier may own code the other imports; cmd/agentctl already depends on
15 internal/common packages, and this adds one internal package and zero
external ones (+14 KB on a 42.7 MB binary).

The quoting helpers deliberately lose their //go:build windows tag: they are
pure string transformations, and tagging them meant loginpty's copy was
compiled by no CI job at all (the Windows job tests a package allowlist that
did not include it), which is how the drift went unnoticed. Their tests now
run on every runner, and the shared package is added to the Windows job so
the ConPTY double-close guard is still exercised natively.

Signatures reconcile on uint16 (what pty.Winsize and PtyHandle.Resize
already take). This is lossless: the resize path widens from uint16, and
DefaultCols/DefaultRows are clamped before use, so no caller was ever out of
range.

Tests add a round-trip property check that decodes buildCmdLine output with
an independent CommandLineToArgvW implementation, verifying the escaping
against its actual contract rather than hand-computed goldens.
