--- apps/backend/internal/agent/loginpty/pty_handle.go
@@ -1,10 +1,0 @@
-package loginpty
-
-import "io"
-
-// ptyHandle abstracts PTY operations across Unix and Windows.
-// Unix wraps creack/pty (*os.File); Windows wraps ConPTY.
-type ptyHandle interface {
-	io.ReadWriteCloser
-	Resize(cols, rows uint16) error
-}
