+++ apps/backend/internal/agentctl/server/process/process_signal_unix.go
@@ -6,6 +6,8 @@ import (
 	"os"
 	"os/exec"
 	"syscall"
+
+	"github.com/kandev/kandev/internal/common/ptyexec"
 )
 
 // terminateProcess sends SIGTERM to the process for graceful shutdown.
@@ -15,7 +17,7 @@ func terminateProcess(p *os.Process) error {
 
 // waitPtyProcess waits for the PTY process to exit and returns exit info.
 // On Unix, uses cmd.Wait() which inspects WaitStatus for signal information.
-func waitPtyProcess(cmd *exec.Cmd, _ PtyHandle) (exitCode int, signalName string, err error) {
+func waitPtyProcess(cmd *exec.Cmd, _ ptyexec.PtyHandle) (exitCode int, signalName string, err error) {
 	err = cmd.Wait()
 	if err == nil {
 		return 0, "", nil
